Rename some things
* HandRolledParser is now RecursiveDescentParser
* AbstractParser is now AbstractCombinatorParser
* AbstractParser is now the common supertrait of RecursiveDescent and
  AbstractCombinator
